requirements.txt
fastapi
uvicorn
gunicorn
line-bot-sdk>=3.0.0
google-cloud-firestore
google-cloud-storage
google-genai
google-cloud-aiplatform==1.71.0
google-generativeai==0.8.3

python-multipart
googlemaps
python-dotenv
requests
pydantic
aiofiles
pandas
pillow
omdb
flet
